Moved audio mixer class into audio system to make it more clear that it is the thing acting on the audio components
