--- FTS/src/main/cpp/engine/include/audio/mixer.h
@@ -1,13 +1,0 @@
-#pragma once
-
-
-#include <cstdint>
-
-namespace rvr {
-class Mixer {
-public:
-    void Render(float* audioData, int32_t numFrames);
-private:
-    float mixingBuffer_[1920];
-};
-}
